Update prerequisites.md
Created new google forms to collect github usernames, so it doesn't get all muddled with the hackathon responses. I updated the link in the workshop prereqs here. Also updated the text to refer to a slack channel for the workshop (instead of the hackathon), but not sure if we're looking to have a slack channel for this or not, or combine with the hackthon channel?